Update: Spark, Gnosis + fix last PR naming
