+++ .vscode/launch.json
@@ -217,6 +217,7 @@
         "morpho-aavev3",
         "morpho-compound",
         "mstable",
+        "mu-exchange",
         "multichain",
         "nemesis-dao",
         "nexus-mutual",
@@ -265,7 +266,6 @@
         "rocket-pool",
         "rook",
         "scream",
-        "sdai",
         "seamless-protocol",
         "sentiment",
         "set-protocol",
