+++ src/adapters/mu-exchange/index.ts
@@ -3,7 +3,7 @@ import type { Adapter } from '@lib/adapter'
 import * as gnosis from './gnosis'
 
 const adapter: Adapter = {
-  id: 'sdai',
+  id: 'mu-exchange',
   gnosis: gnosis,
 }
